PUT's are now id safe

+++ src/main/java/io/github/soydivision/book/Book.java
@@ -44,7 +44,7 @@ public class Book {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
